+++ dp_conceptual_search/ons/conceptual/client/__init__.py
@@ -1,0 +1,2 @@
+from .fasttext_client import FastTextClientService
+from .conceptual_search_engine import ConceptualSearchEngine
